PR 1 (#1570) was squash-merged into main. PR 2 was branched from PR 1
before it merged, so the shared files (__init__.py, fail_open_alert.py,
test_fail_open_observability.py) show up as add/add conflicts even
though the base content is identical.

Kept PR 2's versions because they are strict supersets:
  __init__.py                       — adds notify_customer_fail_open export
  fail_open_alert.py                — adds _also_notify_customer wire-in
  test_fail_open_observability.py   — adds isolation fixture

19/19 fail-open tests still pass after resolution.

Refs #1571

+++ apps/api/app/modules/guard/models.py
@@ -647,7 +647,7 @@ class GuardApprovalRequest(Base):
     decided_by_user_id = Column(String(255), nullable=True)
     decided_reason     = Column(Text, nullable=True)
     decided_at         = Column(DateTime(timezone=True), nullable=True)
-    latency_ms         = Column(Integer, nullable=True)
+    latency_ms         = Column(BigInteger, nullable=True)
 
     created_at = Column(DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
     timeout_at = Column(DateTime(timezone=True), nullable=False)
